Following changes are done. -Utility functions moved to another file name MeterUtils, this file will contain all of the helper functions -Contact name is get against the received message number.
